# Original data should not be corrupted on a delete