… benchmark/benchmark_delete_test.go